# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Social Media API",
    "version": "1.0.0",
    "description": "Professional REST API documentation for the Social Media Backend. Supports authentication, posts, comments, stories, notifications, real-time messaging, and more.",
    "contact": {
      "name": "API Support",
      "email": "support@socialmedia.dev"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current Environment API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "أدخل رمز JWT: Bearer <token>"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/posts/{postId}/comments": {
      "post": {
        "summary": "Add a text comment to a post",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "parentCommentId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added"
          }
        }
      },
      "get": {
        "summary": "Get comments for a post",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments"
          }
        }
      }
    },
...
  }
}